requirements.txt
Flask-RESTful
flask
Flask-Session
mysql-connector-python
flask_cors
python-dotenv
PyJWT
python-engineio

